fix(auto-reply): isolate result contracts
