+++ src/auto-reply/reply/agent-runner-result-complete.ts
@@ -17,8 +17,8 @@ import {
   resolveSourceReplyPolicy,
 } from "./agent-runner-core.js";
 import { normalizeAssistantFinalDeliveryText } from "./agent-runner-core.js";
-import type { accountReplyAgentRun } from "./agent-runner-result.js";
-import type { FinalizeReplyAgentRunInput } from "./agent-runner-result.js";
+import type { accountReplyAgentRun } from "./agent-runner-result-accounting.js";
+import type { FinalizeReplyAgentRunInput } from "./agent-runner-result.types.js";
 import {
   accumulateSessionUsageFromTranscript,
   buildInlineRawTracePayload,
