+++ src/auto-reply/reply/agent-runner-result-payloads.ts
@@ -35,7 +35,8 @@ import {
   hasSessionRelatedCronJobs,
   hasUnbackedReminderCommitment,
 } from "./agent-runner-reminder-guard.js";
-import type { accountReplyAgentRun, FinalizeReplyAgentRunInput } from "./agent-runner-result.js";
+import type { accountReplyAgentRun } from "./agent-runner-result-accounting.js";
+import type { FinalizeReplyAgentRunInput } from "./agent-runner-result.types.js";
 import { resolveResponseUsageLine } from "./agent-runner-usage-line.js";
 import { attachMcpAppChannelAction } from "./mcp-app-channel-action.js";
 import { normalizeReplyPayload } from "./normalize-reply.js";
